Add full support for backend failover.
This works nearly identical to twemproxy: there's an error limit that
must be triggered for a given backend, and once it goes over that error
limit, the backend goes into a "cooloff" period in which it's removed
from the distributor, and then is allowed back after the cooloff period
expires.

Backend "cooloff" is configurable on a per-listener basis.  We also have
an integration test to exercise the feature in a basic way.  By default,
the feature is enabled: 5 error limit to trigger, 10 second cooloff
period.

There's also a lot of cleanup work in here: documentation, formatting
and linter fixes.  Sorry for the tangled changes.