Add support for backend timeouts.
Add the ability to define a timeout on operations to backends.  This
does not affect the "frontend" of Synchrotron at all: a client could
connect and Slowloris us and this wouldn't do anything for that.  By
virtue of batching requests to backends, this means that the timeout
itself could technically be a result of batch sizing: if you submit a
single batch of 128 items as a client, even if the first result comes
back before the timeout, the batch itself might partially fail for
whatever messages haven't gotten responses yet.

By default, we've set the timeout at 1000ms.