protocol: fixups to redis to prepare for shadowing
We needed some machinery to send commands to the server to prevent
server replies, and `from_inline` only worked for non-RESP commands, so
we made it more legit.

Also created a helper future type to wrap futures that deal with
utilizing a backend connection.  This should help us eliminate a lot of
type constraints on the processor.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>