processor: add support for "noreply" mode in processor
We've now baked in support for "noreply" mode in processors.  This
allows backends to choose whether or not they want to run a request and
get back a response or not.  This is a core part of traffic shadowing in
that we want to send all requests to a shadow backend but don't actually
_need_ the responses.

We've also switched to using ProcessFuture and have dropped a _lot_ type
constraints on the processor - woohoo!

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>