Add command filtering for Redis.
This adds the ability to disconnect a client if they use unsupported
commands.  This is required because some commands that have multiple
arguments that would otherwise be sharded -- like BITOP and the N number
of source keys it takes -- can only ever be routed to a single Redis
node.  Since that node would only be valid to get the value of the first
key in the key, those command would produce unexpected or incorrect
results.